investigate: move renderer to investigate/flowchart, drop dev script
The mermaid renderer is only used by investigate, so nest it there. Rename
the package mermaidascii -> flowchart (single word, names the domain
object) and RenderFlowchart -> Render to avoid stuttering at call sites.

Remove the render-mermaid preview script, its mise task, and the example
diagrams — dev-only scaffolding not worth carrying in the repo.
