--- diagrams/2-branch.mmd
@@ -1,5 +1,0 @@
-flowchart LR
-  R[Read config] --> V{Valid?}
-  V -->|yes| START[Start server]
-  V -->|no| ERR[Log error]
-  ERR --> EXIT[Exit 1]
