--- diagrams/5-multi-root.mmd
@@ -1,5 +1,0 @@
-flowchart LR
-  CLI[CLI invocation] --> H[Hook handler]
-  H --> LOG[Write log]
-  GUI[IDE plugin] --> H2[Hook handler v2]
-  H2 --> LOG2[Write log]
