--- diagrams/6-deep-subgraph.mmd
@@ -1,10 +1,0 @@
-flowchart LR
-  subgraph client [Client side]
-    U[User runs cmd] --> A[Agent spawns hook]
-  end
-  A -->|stdin pipe| B["entire hooks<br/>session-start"]
-  B --> C[io.ReadAll blocks on EOF]
-  C -->|EOF| OK[Parse + dispatch]
-  C -->|no EOF| STUCK[Hang]
-  STUCK -->|killed| FAIL[No session logged]
-  FAIL -->|user retries| U
