--- diagrams/7-fallback-sequence.mmd
@@ -1,5 +1,0 @@
-sequenceDiagram
-  participant A as Agent
-  participant H as Hook
-  A->>H: session-start event
-  H-->>A: ack
